Update README.md
Not sure if it was intended or not but after installing `version 1.0.1` I found that the polybar module had to be changed from `exec = bar-protonmail` to `exec = bar_protonmail`. I'm not sure if it needed the same thing in the waybar module or not as I do not have that installed.