Update README.md
Not sure if it was intended or not but after installing version 1.0.1 I found that the `bar-protonmail` had to be changed to `bar_protonmail` for the `auth` command, in the `polybar module` section and in the `script arguments` section. I'm not sure if it needs the same thing in the waybar module or not as I do not have that installed, but I would think it might.